Separated the core of main to a separate function.
main() in api.py is basically the same as the main() in cli.py with the only differences being how it sets up and what it does afterwards. This allows for all current and future uses to use the same thing without copy and paste, and makes it easier to add/change behavior in the future.