Created HTML page.
Removed usage of Node.js as it is not for static web pages.
Even Split does not account for multiple inputs yet.
The settings page is completely non-functional besides the Light / Dark Mode button.
All valid input should work, some invalid may still not be caught.
The first line (import) of `ConveyorNode.js` needs to be removed transpiling for it to work.