Broke up large Vite index-[hash].js
With the normal imports, Graphviz and parts of Mermaid where getting put into the main js file, making it relatively large.
Using dynamic imports allowed them to be put in separate chunks.

Additionally, `vue-mermaid-string` has a lot of resulting chunks. While most frankly aren't needed, I'm not aware of a way to exclude them without replacing it.
While it won't help with (the non-issue of) network usage, the submodules are sorted into a folder to help it feel more organized.