Moved switched to multiset math and moved pure math functions to dedicated file
In cases with a multiple of the same Source or Target, the permutations would count / run duplicates.
When calculating based on a multiset, the only unique permutations are counted / ran.
Since permutations increase O(n!), it's very important to reduce duplicate permutations.

Since `splitRatio.ts` had large portion replaced with this commit, I also did some touch up work in there as well.

I didn't allow Prettier to run in `ConveyorNode.ts` as I didn't want all of that in this commit.
It will get its facelift soon as the `primeFactorization` is intended as part of a rework of the main algorithm.
Mainly, knowing the prime factorization should allow splitting the bigger primes first, like mentioned in #6.