Reset node ids for newly calculated graphs
This makes it so each new graph's ids start at `0` rather than wherever the previous graph left off.

It was going to be done in the `main` and `main_split` functions, but it ended up not doing anything thanks to vue's reactivity proxies.
This could be worked around, but that's more work than it's worth.